requirements.txt
pygame== 2.6.1
psycopg2-binary==2.9.7
pydub
simpleaudio
Pillow